+++ src/services/entity.service.mts
@@ -11,8 +11,6 @@ import {
 import dayjs, { Dayjs } from "dayjs";
 
 import {
-  ALL_DOMAINS,
-  ANY_ENTITY,
   EditLabelOptions,
   ENTITY_REGISTRY_UPDATED,
   ENTITY_STATE,
@@ -21,9 +19,9 @@ import {
   EntityHistoryResult,
   EntityRegistryItem,
   HassEntityManager,
-  PICK_ENTITY,
   TMasterState,
-} from "../helpers/index.mts";
+} from "../index.mts";
+import { ALL_DOMAINS, ANY_ENTITY, PICK_ENTITY } from "../user.mts";
 
 const MAX_ATTEMPTS = 10;
 const RECENT = 5;
