+++ src/testing/area.spec.mts
@@ -1,8 +1,8 @@
 import { sleep } from "@digital-alchemy/core";
 
-import { TAreaId } from "../dynamic.mts";
 import { AREA_REGISTRY_UPDATED, AreaDetails } from "../helpers/index.mts";
 import { hassTestRunner, INTERNAL_MESSAGE } from "../mock_assistant/index.mts";
+import { TAreaId } from "../user.mts";
 
 describe("Area", () => {
   const EXAMPLE_AREA = {
