+++ src/testing/entity.spec.mts
@@ -1,7 +1,8 @@
 import { sleep } from "@digital-alchemy/core";
 
-import { ANY_ENTITY, ENTITY_STATE } from "../helpers/index.mts";
+import { ENTITY_STATE } from "../index.mts";
 import { hassTestRunner } from "../mock_assistant/index.mts";
+import { ANY_ENTITY } from "../user.mts";
 
 describe("Entity", () => {
   afterEach(async () => {
